display: complete alpha overflow fix, cast everything to u32

+++ src/display.nim
@@ -117,9 +117,9 @@ proc renderTextLine(fb: ptr UncheckedArray[uint16], font: ptr stbtt_fontinfo,
             var bgR, bgG, bgB: uint8
             fromRGB565(bgPixel, bgR, bgG, bgB)
             let inv = 255'u32 - uint32(alpha)
-            let newR = uint8((uint16(cr) * uint16(alpha) + uint16(bgR) * inv) div 255)
-            let newG = uint8((uint16(cg) * uint16(alpha) + uint16(bgG) * inv) div 255)
-            let newB = uint8((uint16(cb) * uint16(alpha) + uint16(bgB) * inv) div 255)
+            let newR = uint8((uint32(cr) * uint32(alpha) + uint32(bgR) * inv) div 255)
+            let newG = uint8((uint32(cg) * uint32(alpha) + uint32(bgG) * inv) div 255)
+            let newB = uint8((uint32(cb) * uint32(alpha) + uint32(bgB) * inv) div 255)
             fb[fbBase + py] = toRGB565(newR, newG, newB)
       {.pop.}
